Masked PointOfInterestService behind an interface
